Added initial version finder
This commit adds the `version-finder` tool which looks for text patterns
in git repo to find in which versions and commits certain features were
introduced.

The tool will iterate over a configurable list of tags from newest to
oldest, checkout the tag and perform text searches for all patterns.
It will record the earliest tag where the pattern appeared.

In the second phase, the tags discovered in the first phase are used to
perform bisections between the first version with a text pattern and
the one before to find the exact commit where it first appears.

All of the gathered information is stored in a YAML file to be used by
other tools to generate or annotate documentation.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>