Add feature tag generation tool
This commit adds the feature tag generation tool which takes the feature
version and commit information generated with the version-finder tool
and generates markdown version tags between designated markers.

These tags are the exact same style as before, having the icon and
version number visible and linking to the linux kernel github to the
exact commit where the feature was introduced.

This is going to safe a lot of manual labor and potential mistakes

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>