refactored initial map concept page
Went over the original page and refactored it, deleted some redundant
parts we already had in other locations and linked to them instead.
Move a number of the examples to out to those specific pages since
we did not yet have those.

The BTF defined map still needs a bit of expansion to explain what the
different map field names mean, we do not have that info anywhere else.

The last two paragraphs about pinning are not map specific, they should
be moved to their own concept page and be slightly reworked to be more
generically worded.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>