Added custom stylesheet to widen navigation sidebar
The navigation for this project is somewhat deep with up to 4 or 5
levels at the moment. In addition to that, a lot of the kernel
constants included in the titles make them take up more width.

On larger monitors the Material theme adds padding/margin on the sides
to center the content. The style changes introduced here allow us to
use more of this before unused space to widen the navigation sidebar
a bit. On tables or mobile the dynamically displayed overlay is also
slightly bigger.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>