Add "since" information for helpers, specifically bpf_{g,s}etsockopt
Normally when helper functions are added they are added to all program
types that exist at the time. When a new program type is added, all
existing helper functions that should be usable for that program type
are added in the same kernel version. However, it sometimes happens
that a helper function is added to a program type in a later kernel
version in which case both the version number on the program type and
the helper function page are incorrect.

This is the case for bpf_{g,s}etsockopt. This commit adds the correct
version numbers to the data files, updates the reference generation
tool and updates all effected references.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>