Docker commands now use the UID:GID of the user
Modified our docker image and run commands to use the UID:GID of the
current user inside the docker container. This means that any files
created inside the container will have the same UID:GID as the user
outside the container and we don't end up with files owned by root.