Add kfunc-gen tool to generate kfunc documentation
This tool looks at the data/kfuncs.yaml file and uses the names of the
kfuncs to find their function signatures in the vmlinux file. It then
translates the BTF into a C function signature as they would appear in
the kernel source code.

The VMLinux blob was extracted from a kernel build using the
cilium/ci-kernels repository. A little customization was required to
compile in all the kfunc signatures. Instructions for updating the blob
have been included.

The tool also generated the list of program types that can make use of
the kfunc.

Lastly the tool generates notices and warnings based on the flags given
to the kfuncs. These are placed below the function signature.