tools: Add spellcheck tool
This commit adds a new spellcheck tool. The purpose of which is to be
our first CI tooling to check for spelling errors. Due to the use of
markdown extensions and code generation, it is difficult to use ready
made tools. What our custom tool does is render the markdown files to
HTML, parse the HTML with custom logic and exceptions, then feed the
text into `aspell` (GNU spell checker) to check for spelling errors,
and then do a best effort to find the location of the spelling mistake
in the markdown files.

Aspell allows to the addition of a personal word list, which will be
added in a followup commit. In addition, the tool excludes code blocks,
text between <nospell></nospell> tags, generated output, and html which
is not from the markdown files.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>