Makefile,tools: Add pre-built Docker image
The project is seeing more activity which is great, but it also means
we have to run CI/CD jobs more often. Up until now CI has had to
rebuild the Docker image from scratch every time, since it doesn't
enjoy the same caching benefits as a local build. This is a waste of
resources and time.

This commit changes a few things. First, the image has a feature where
it will switch the the same UID:GID as the current user so that files
created in the container on the mounted volume are owned by the user
and not root. This logic used to be done at image creation time, but
now has been converted to a script that runs at container start time.

This allows us to build the image once and then use it for CI/CD jobs
as well as locally. The container-build target in the makefile will
now build the image locally and tag it so that it can be pushed to
the github docker registry by those with permissions. This should not
occur often, so for now this is still a manual step.

All makefile targets that need the image should now pull the pre-built
image if it is not already present. If for some reason the image is
unavailable, it can still be recreated locally with the build target.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>