tools/spellcheck: Remove project root from spellcheck output
The spellcheck will output its best guess for the location of misspelled
words in the markdown files. This is useful for the user to quickly find
and fix the misspelled words. However, the output also includes the
project root, which is likely different for every system including the
host vs in a docker container. By removing the project root from the
output, we get a relative path which most terminals can use to open the
file as long as the working directory is the project root.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>