helper-functions: Add 'since' data to bpf_sk_storage_{get,delete}
The bpf_sk_storage_{get,delete} helper functions were added in kernel
in version v5.2, and tracing programs in v5.5 yet it wasn't until v5.11
that the helper functions were allowed to be used in tracing programs.

This had not been reflected in the docs so far, this commit fixes that.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>