Update `bpf_get[_ns]_current_pid_tgid` helper function data
Since v6.10, the `bpf_get[_ns]_current_pid_tgid` helper functions can be
used in all program types. This was not reflected in the docs.

Since we want to reflect the version in the references, we are keeping
the existing program associations and adding a new mention of the
helpers in the "base" group with the `since` attribute. The intended
purpose is so all programs that could use them before are not annotated
with the `since` info, but all programs where support is new do have the
version info.

However, turns out the group flattening logic did not respect the
hierarchy of the groups, so the `since` attribute was always shown.
So the tool was updated, groups are resolved first, if a more specific
group or more specific program type declares it as well, the newer
declaration is used. This results in the intended behavior.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>