tools: Make libbpf based tools use a stable reference
We would like to be able to run assert in CI that running
`make generate-docs` will not change the output of the documentation.
This avoids issues where someone manually modifies bits of the docs
that are supposed to be generated.

However, we have a few tools that use files from libbpf as input, which
so far we have been pulling from the `master` branch which is always
in flux. This means that the output of the tools can change between
runs.

So, lets use a pinned commit hash for the libbpf reference. We will
manually have to update it periodically to pull in changes from libbpf.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>