Fix git committers plugin and its cache
By default, the repo gets checked out with a depth of 1 which causes the
git committers plugin to think all commits are authored and updated
on the current date. This causes it to not be able to reuse cache
and makes it emit the wrong dates in the page headers.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>