kfunc-gen: Kernel v6.13 updates
Updated the vmlinux to be that of a v6.13 kernel. Updated the
instructions to add KCONFIG options that new kfuncs are gated behind.

Added all new kfuncs to the kfuncs data file. And changed the format
of the datafile so we can now associate `since` and `until` values
with the kfuncs -> programType relation. The tool has been updated to
output version tags besides the kfunc -> program and program -> kfunc
reference lists.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>