Add custom hook to allow links in code blocks
While documenting struct ops I felt the need to be able to add link in
the code blocks. This was not possible out of the box, so I wrote a
crude MkDocs hook that detects the markdown link pattern (`[text](url)`)
and replaces it with a proper HTML link.

For the struct ops pages it will allow me to link from the struct
definition to the section containing more information about the field.
But I also imagine this could be very useful in examples where we can
now link to the definition of a function.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>