Highlight sections when navigating to them
This commit adds a hook that adds a `<section>...</section>` around
every logical sections based on the headings in the markdown file.
This allows us to add CSS for whole sections at any level.

Secondly, add some CSS which adds a yellow stripe on the left side of
the section that is "targeted". A section is targeted when the browser
just navigated to it via a link. This stripe makes it easy to see where
the section of interest starts and ends. We also briefly make the
background a blue-ish color which fades out in 2 seconds. This makes
it pop and draws the users eye to the section of interest.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>