Switch to monospace font and enable code annotations
We have been using a non-monospace font for code blocks. This messes up
any deliberate alignment of text in code blocks. So switching to a
monospace font for code blocks.

We also enable code annotations, which allows us to write comments in
code blocks which will be turned into marks that can be expanded. This
is neat for adding some info in situations where the extra info is to
small to warrant a whole section.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>