Add pages for specific struct_ops.
`BPF_PROG_TYPE_STRUCT_OPS` programs are a bit special in the sense that
even though they are a single program type, they can be used to a wide
variety of function callbacks in the kernel.

Every "struct_ops" that can be implemented via BPF has different
semantics for which fields are available, how and when they are called,
and what you can do with them.

It therefore makes sense to have a separate page for each struct_ops
implementation, to document the specifics of that implementation.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>